+++ include/pos.h
@@ -1,13 +1,12 @@
 #pragma once
 
 #include <stdbool.h>
-#include <stdint.h>
 
 #include "dir.h"
 
 typedef struct {
-    size_t row;
-    size_t col;
+    int row;
+    int col;
 } Pos;
 
 Pos pos_new_invalid(void);
